refactor: add useAuthUser hook
